🔧 Add plugins import and styles for modal z-index
This commit adds an import statement for the "plugins" file in the main.scss. It also introduces a new file called "plugins.sass" which contains styles for the .wcm-modal class with a z-index of 10000.

🎨 Adjust width and max-width in Waiting.vue

In the Waiting.vue component, this commit adjusts the width property of the .marquee class to be 100% instead of 100vw. The max-width property is also set to 100%.

📐 Update flexbox styles in burrito.vue

The burrito.vue layout now has additional flexbox styles. The flex-direction is set to column, and there are new properties added: max-width is set to 100vw and overflow is set to clip.